refactor: import logic to seperate component
